refactor : theme 를 recoil에서 types, utils로 분리
