+++ src/app/create/CreateTheme.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import BottomButton from '../../component/BottomButton';
-import { defaultState } from '@/recoil/newtheme/default';
+import { defaultState } from '@/utils/theme/default';
 import Image from 'next/image';
 
 interface CreateThemeProps {
