+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import { defaultState } from '@/recoil/newtheme/default';
+import { defaultState } from '@/utils/theme/default';
 import Image from 'next/image';
 import Link from 'next/link';
