+++ src/component/HeaderLayout/DropListItem.tsx
@@ -2,9 +2,9 @@ import { useRecoilState } from 'recoil';
 
 import { Check } from '../Icon/Drawer';
 import { drawerState } from '@/recoil/drawer';
-import { defaultState } from '@/recoil/newtheme/default';
 import { themeState } from '@/recoil/newtheme/theme';
 import { drawerBoard } from '@/types/drawerBoard';
+import { defaultState } from '@/utils/theme/default';
 import { usePathname, useRouter } from 'next/navigation';
 
 interface Props {
