+++ src/component/HeaderLayout/TitleHeader.tsx
@@ -6,10 +6,10 @@ import Close from '../Icon/Close';
 import { HamburgerMenu } from '../Icon/Header';
 import useModal from '@/hooks/useModal';
 import { useScroll } from '@/hooks/useScroll';
-import { defaultState } from '@/recoil/newtheme/default';
 import { drawerState, themeState, titleState } from '@/recoil/state';
 import { confirm } from '@/utils/confirm';
 import { drawerOpen } from '@/utils/drawerOpen';
+import { defaultState } from '@/utils/theme/default';
 import { usePathname, useRouter } from 'next/navigation';
 
 export default function TitleHeader() {
